Ensure nodes have stable instances in the eyes of stetho
Summary: Stetho currently has a bit of a flawed model for identifying a node. If a node's underlying instance changes then it will receive a new ID. This causes the node to be added and removed from chrome loosing all state. This is bad as if this node had children it will loose the state that it was expanded and thus the whole element tree will collapse making it very hard to edit multiple properties. This diff works around this by mapping component keys to wrapper instances which we then return to stetho.

Reviewed By: passy

Differential Revision: D4566543

fbshipit-source-id: 852f58212378120aac2cb1fe56450b208eb6c36e