Create our own global key which handles nested trees
Summary: We were in some cases generating the same node it for two different nodes in a tree. I think it is due to global key not being meant to be global across nested trees. We can get around this by constructing our own global key based on the position a node in the tree.

Reviewed By: IanChilds

Differential Revision: D4589516

fbshipit-source-id: 1924b9c1f4a134e5b0b0043fdae792786c18f1c2