TransitionKeySet handles all transitions for a given key
Summary: TransitionKeySet is a class grouping all Transitions for a given key and tracking the global state of all the values the transitions for that key wants to animate. The only responsability of the class is to start, stop or resume transitions for a key keeping track of its state and making the right decision on which transition to start or resume.

Reviewed By: muraziz

Differential Revision: D4590816

fbshipit-source-id: 4b459f28b104e1a781d275ed58a01ed38da2e096