Only set the text if it actually changed
Summary: This prevents the cursor from jumping to the end/start of the text when typing, if the EditText's text is bound to some State. Not sure if this is complete, but I'm also not sure how to improve it. Let me know what you think. More context in https://www.facebook.com/groups/1430200360634661/permalink/1789624414692252/

Differential Revision: D4571383

fbshipit-source-id: 05adcc327b49b7bb20a053be6d29f81149b1bd5b