Lazy State Update in SectionComponent
Summary:
Lazy state update will be used in FeedStorySection. But lazy state update is currently not working in SectionComponent.

In `applyNewChangeSet()` after the first proper state update, the previous lazily updated states will be reflected when the new SectionTree is created.

However, when we re-create the SectionTree, new StateUpdate from the lazy updates will be added to the `mPendingStateUpdates`. This causes the `mPendingStateUpdates.size() == 0` check to fail. This cycle will keep going on which means that we will never get out of the while loop.

The proposed solution here is to check for the present of any non-lazy state update in the pending list. And we can exit the loop if there isn't any.

Reviewed By: pasqualeanatriello

Differential Revision: D4655361

fbshipit-source-id: ac75b640e3748f5547c22ebb788f071a1b26b252