Enable incrementalmount option for ComponentTree by default and reflect changes accordingly
Summary: Before opensourcing Components we would like to remove optional features not to confuse developers. Incremental mount is well tested and enabled in NewsFeed, and we are turning it on by default. Reflecting other usages accordingly, i.e. remove where it explicitly enables and add a line (`.incrementalMount(false)`) where nothing is specified.

Reviewed By: IanChilds

Differential Revision: D4681438

fbshipit-source-id: 4db61d81467e69d2f9ec3e02697936e9b1958e52