Use latest SoLoader
Summary:
The old version of SoLoader only loads 32 bit natives.  This is not acceptable if we are using 64 bit fresco libraries (ie it causes crashes).

This diff introduces an aar for the new soloader.  **I DO NOT LIKE THIS.**  We should be fetching this, as it is open source and hosted on bintray, but it seems internal commits are not sync'ed and who knows when bintray will be updated.  This is what is wrong with `/lib`.  As we hack internal libs better and better, we will be forced to do this to more libraries.  Can't we have some sort of automatic maven push?  </rant>

Reviewed By: passy

Differential Revision: D4712799

fbshipit-source-id: e696a1e6bd391ce201fc99c5a53581dfe7a5dfcd