Remove annotation processor note spew and filtering script
Summary:
After fixing the corruption of error messages by Ultralight, Buck's message
sorting is kicking in and sorting the root cause error to the front of the list.
At this time there will still be missing type errors for generated types, but the
root error(s) will be the first one(s) listed.

We could look at outright removing missing type errors if there are other ones present,
but that feels like it could also be confusing. Developers are used to compiler errors
getting less and less relevant the further down the list they get, so it's probably
best to err on the side of reporting them all but sorting them reasonably.

Reviewed By: emilsjolander

Differential Revision: D4686467

fbshipit-source-id: f160f89269c44a4f0e926fc4093acd6ec5ac8890