Move custom shadows to //libraries
Summary:
I'm cheating here by putting LayoutDirectionTest on the powermock list even
though it doesn't actually use it in a way that would normally require it.
However, due to it using a custom Robolectric config it will conflict with the
Yoga loader which is why we need to run it separately.

Reviewed By: rspencer01

Differential Revision: D4737705

fbshipit-source-id: af3e1ad2917ac0f59a2ae8ed5995990ab516a7cc