Fix 0 height ComponentView recycling issue
Summary:
There was a report that some components are not rendered in Pages Manager app (https://fb.facebook.com/groups/1430200360634661/permalink/1809509549370405/). In some cases a component existed in the tree hierarchy, but its height was 0. Later after talking with people I found that this is a well-known issue that is not yet solved. For some reason multi-row in NewsFeed needs to set the height of the component to 0 when they don't have enough data to render component and when that component is recycled and re-used elsewhere we end up with this issue. Recycled component is reused and measurement of that component terminates early because we are checking for height==0 or width==0 before deciding to measure it.
There was another approach taken by emilsjolander (D4611416) but it was reverted (D4631574) because for that to work every subclass of `ComponentView` needs to make sure to set that measured flag which is not obvious thing to do when subclassing `ComponentView`.
This diff makes sure to terminate early from measurement only when *both* width *and* height are 0.

This fixes the pages manager issue. Feel free to suggest some other approach if you find a better solution.

Reviewed By: marco-cova

Differential Revision: D4779019

fbshipit-source-id: bc5b80840b3f4c951b434ccba39cbcc1e99f7ed1