Fix sticky header layout issue when it was passing left of parent to its child.
Summary: We found that the layout logic of sticky header had a bug when it was incorporating the parent's left/top into its child, thereby positioning sticky header in incorrect location. This issue becomes visually obvious when `RecyclerViewWrapper` is laid out not on (0,0) top-left corner of its parent, but some different values.

Reviewed By: pasqualeanatriello

Differential Revision: D4794748

fbshipit-source-id: 7896715f32c64788946c1dc8217c5072e77858e4