Fix touch event handler given by '.touchHandler()' not being fired in RecyclerSpec
Summary:
When we pass `EventHandler` to `RecyclerSpec`'s `.touchHandler()`, it is never triggered. That is because `RecyclerSpec` mounts `RecyclerViewWrapper` which contains `RecyclerView` as a child. And `RecyclerView` always consume touch, thereby not propagating touch event to its parent where user-provided EventHandler is located.

This diff makes sure that when we pass `EventHandler` to `RecyclerSpec` it goes directly to `RecyclerView` instead of its wrapper parent.

Reviewed By: IanChilds

Differential Revision: D4788856

fbshipit-source-id: 907162e08d141d11c4e49507ef6f2482e29c1176