Add PendingNode to specify dataflow nodes that need to be resolved at runtime
Summary:
Depends on D4737608

Since some nodes relating to Views can only be resolved at runtime (e.g. after a binding has been declared), we need to add specs when declaring a binding, and then resolve them when the view is about to mount.

The way this works in this diff is that a GraphBinding can also collect PendingBinding (a package private concept) that are between a ValueNode and a PendingNode. Example PendingNode are for things like a view property, or something based on a view property like a constant offset from the initial render position.

Before a binding is activated, the framework must call resolveSpecs on it and return a ValueNode for each PendingNode the binding has. The code crashes if activate is called while there are still unresolved specs. The GraphBinding code handles then adding the bindings between the newly resolved nodes and the nodes they should be bound to as per the PendingBinding.

Reviewed By: emilsjolander

Differential Revision: D4769852

fbshipit-source-id: 24d870b8a45bfed2a897258a5c799483fb7b894a