Fix canMountIncrementally
Summary: Previously we were just storing the index of the layout output, rather than the id. This led to an issue where the index may not be present if it is moved (due to mount diffing). Therefore we should store the id in the same way that the mount item map does.

Reviewed By: pasqualeanatriello

Differential Revision: D4802123

fbshipit-source-id: af96972d7b92df25b014abe017b079738b45b075