Fix the sticky header measurement
Summary: When measuring sticky header we were passing parents measurements as is. But if there is no sticky header `RecyclerViewWrapper` was creating a view that's GONE and has same height as wrapper. Even though this was not affecting when sticky header was present, we don't want to have non-0 height sticky header when it is not used. That is especially apparent when one uses Stetho and clicking Wrapper was always hitting sticky header instead of the child item of RecyclerView.

Reviewed By: pasqualeanatriello

Differential Revision: D4810141

fbshipit-source-id: ef740214f97b953de30bcd73ffdfeb28841d598c