Treeprops ThreadSafety
Summary:
t16148778 shows that when background layout is enabled, the TreeProps map is being copied and cleared concurrently.
I'm guessing the reason is that a new layout pass (which can be on a different thread than the prev layout pass) would copy whatever treeprops were in the context.
However, the root in the new layout pass should actually have null treeprops, with the exception of nested trees. If the tree is nested, using the pending props is safe, because pending props are never cleared.

Reviewed By: pasqualeanatriello

Differential Revision: D4786967

fbshipit-source-id: ebc5c57952473920c2cfe316b50c0ee44199557c