Replace Container with Row & Column
Summary:
It is not obvious whether the default direction for a Container is ROW or COLUMN. This is not a problem unique to Components, e.g. what is the default for LinearLayout? Android provides a lint error in xml files which suggest always specifying the orientation of a LinearLayout. This is good but I think we can do better.

See https://fb.facebook.com/groups/1785819565004653/permalink/1842819652637977/

Reviewed By: IanChilds

Differential Revision: D4699537

fbshipit-source-id: 27a6ecdd3a7c825d8918079add9b64e0448bbcd4