Remove fbcore dependency
Summary:
Add a new `litho-common` module which we can add other stuff to, like possibly
an `ArraySet`. We need to remove `fbcore` as it's an AAR (and generally
way overblown for a single class) which cannot be included in an AP.

The dependency is still in the repo. I'll remove it with the next patch.

Reviewed By: IanChilds

Differential Revision: D4819360

fbshipit-source-id: 88c568421972512c6a9f5454a685e9623069c187