Add ability for nodes and bindings to finish
Summary:
Certain nodes can finish (spring and timing, for example), and they can tell us when a binding is finished. The basic idea is that the concept of finished can propagate through the graph just like a value: a node can be finished if its inputs are also finished, and if it's node-specific implementation will keep its value from changing.

I'm using this to implement things like Animated.sequence/parallel in a later diff.

Reviewed By: emilsjolander

Differential Revision: D4763827

fbshipit-source-id: 1877d6b48db963c9b1dad9aea857288b53cbc7ed