Add META-INF for processor
Summary:
This is a well-known entry point to detect the processor when this is loaded as
a JAR. I haven't investigated yet if there's a way to tell Buck to pick this up
for artifact generation, but as long as you use this directly with Buck it
is unnecessary anyway.

N.B.: If we had Google Auto, we could use `AutoService` which is super cool and
takes care of updating the package name if it ever changes. One day ...

Reviewed By: IanChilds

Differential Revision: D4826765

fbshipit-source-id: c2bb5db455473a1d9b7041479575b161fe25246b