No scoped componentcontext in mount and bind
Summary: For HostComponents, we don't set a scoped ComponentContext during layout calculation because it would never be used, as we can't update state for them. It's okay to use the MountState's component context that is passed from the ComponentView, which has the same info but it's not scoped to any components.

Reviewed By: IanChilds

Differential Revision: D4795136

fbshipit-source-id: 8e675c4d8770444a9339d8a243cc7a039c73a958