Initial gradle build
Summary:
We're finally here! \o/
This doesn't get us there all the way, but a big step closer.

This will build all the individual library JARs and the widget library. What's
missing right now is a meta-package, the sample app, the tests and ultimately
the CI integration for all of that.

But you can already open this up in Android Studio and build to your heart's
content.

Reviewed By: IanChilds

Differential Revision: D4794818

fbshipit-source-id: 502c64eced41cac257888ab945078c99225a2ae9