Move Yoga to Gradle-friendly location
Summary:
Another Yoga move. Sorry. This one moves the Yoga JNI files and headers and C
files to appropriate subdirectories, that can be picked up by Gradle as an NDK
target. It requires a certain directory structure which is achieved by this
patch.

We should hopefully be able to later remove some of the dependencies that
currently sit in lib/yogajni.

Reviewed By: emilsjolander

Differential Revision: D4843370

fbshipit-source-id: a897ea13c60d0c0b20e49113aa968653c0813758