Fix CardSpec's layout for content and CardClip
Summary:
I was having the most difficult time getting my content to fill the full size of the Card,
which itself was manually sized with .width and .height. The outer layout was the right size,
and the drop shadow was the right size, but the inner content and the clip were not.

Reviewed By: emilsjolander

Differential Revision: D4840448

fbshipit-source-id: 51ff1ccf98d5d7d8a6db703f3d2b745ba5b1d514