Move framework tests to litho-testing
Summary:
I don't like this. But it's a necessary stop gap.

The helpers in `litho-testing` depend on `litho-core`. That means, we cannot
test `litho-core` as it *uses* `litho-testing`. At least not in gradle land,
because we don't have the same granularity there as with Buck.

It's not a disaster, just feels a bit out of place. We may find a better way to
break things down, but for now this seems better than having `litho-core` and
`litho-really-core`.

Reviewed By: IanChilds

Differential Revision: D4843017

fbshipit-source-id: 6877984c696adc74f81565fba2295759d4397e1c