Add setRootAnimated and variants
Summary:
I'm not in love with this tbh, but I haven't thought of a better way. This reverts D4794800 and instead allows setRoot calls to explicitly specify whether a transition animation should happen. This flag then becomes part of the layout state that is mounted.

We'll need to follow up with what to do when setRoot is called without animation while an animation is happening.

Note that this doesn't fix the attached task: in a followup diff, I'll use PersistentState to determine which variant of setRoot to call.

Reviewed By: pasqualeanatriello

Differential Revision: D4835556

fbshipit-source-id: 607f435186e8f31d33f07f4ee083bda94cfed98f