Add NDK build target for Yoga
Summary:
This creates a new gradle sub-project for Yoga.

The sub-project builds the Yoga C lib and JNI bindings in one go. The lib is
built with CMake while the source is still confined to a sub-directory, so we
can sync it with dirsync and keep it as BUCK target the way it was before.

Unfortunately, I didn't manage to build local libraries with CMake (i.e. ones
that run off of an Android device) so we're still shelling out to Buck for this.
I hope we can fix that soon. But hey, unit tests with Gradle!

Reviewed By: emilsjolander

Differential Revision: D4835226

fbshipit-source-id: 94a7f550d098d95a6c1934c00dff781e60f50538