Move framework tests to separate package
Summary:
Another move. I'm truly sorry, I hope this is the last one.

Our framework tests integrate various parts and aren't well isolated with
regards to their dependencies. While Buck provides affordances for dealing with
this on a per-directory basis, Gradle lacks the granularity required for such a
setup.

The `-it` suffix indicates an integration test suite, which is normally
associated with `androidTest` style tests which run on emulators. For now, we
only have Robolectric-based tests, which do however integrate many parts of the
Litho framework.

Reviewed By: IanChilds

Differential Revision: D4851315

fbshipit-source-id: cb0020ac41e57ef81d8a6442a7cc8889333981ad