Update assertj to 2.6
Summary:
This should be fully backwards-compatible. It's the last public release of
AssertJ which doesn't require Java 8 for some parts and also provides java 6
backwards compatibility for Android, which were lacking in the 1.6.1 release we
had previously used.

It adds an important feature that we need: Java6Annotations. These classes
deliberately exclude Java 7 (and 8) references which will break
Robolectric/Android otherwise.

I tried to upgrade to AssertJ 3, but it seems to always pull in
com.java.function.Consumer which is Java 8 only.

Reviewed By: IanChilds

Differential Revision: D4859633

fbshipit-source-id: 1673caad5197da283025698d539507f4f1025867