Fix classloader reuse issue in tests
Summary:
I'm so glad I found this option. This means we're limiting the test classes that
are run per process within this project to one. This means that we'll never
reuse a ClassLoader. It's a bit overkill as there are plenty of tests that could
actually share a ClassLoader and benefit from the speedup, but we don't get that
level of granularity in Gradle. FWIW, this is the same approach that ReactNative
uses. And on multi-core systems the penalty isn't actually that high.

Reviewed By: emilsjolander

Differential Revision: D4867404

fbshipit-source-id: f8142aa518a1c71c1ea3de634f5ef50ea3bc6f4a