Attempt to install with sdkmanager
Summary:
So, the Android Tools folks decided to replace `android` as SDK manager and
replaced it with the ... LOVELY sdkmanager tool which won't even give you a full
list of things it can install. Anyway, it appears that build tools newer than
23.0.2 aren't available on the old one, so we have to use the new one for the
more recent stuff.

Let's see if this works for Circle.

Reviewed By: emilsjolander

Differential Revision: D4876010

fbshipit-source-id: bd307cbedbd1ef0237661ded0d9909f0b601ecda