Add initial Animated end-developer API
Summary:
Adds the Animated class which will be the main interface developers use for defining animations. It wraps classes from previous diffs in this class like `sequence`, `parallel`, `spring`, etc.

The idea behind the different subclasses of AppearingComponents, etc is that we want to provide type safety for the kinds of inputs developers provide when defining an animation. It means that developers will need to explicitly decide whether a component is appearing, changing, etc in their transition.

Reviewed By: emilsjolander

Differential Revision: D4853729

fbshipit-source-id: 83b38e5094bc6f64430012c9fcd2b6e69a7a5809