Move InlineLayoutSpec from litho-core package to litho-testing
Summary: We want people to use `InlineLayoutSpec` only in testing, and not in production. `InlineLayoutSpec` is not designed to be used in prod because it breaks lots of assumptions that framework makes for optimizing the performance. For that reason we are moving it out of core package to testing package.

Reviewed By: passy

Differential Revision: D4875928

fbshipit-source-id: c127e1b971a5f6318d5bf955bd0c0c6f6e7e4b7e