Fix crash with EditText null background
Summary:
In D4841344, we support overriding the background for an
EditTextSpec. This causes a problem when we recycle the EditTextSpec, since,
if no background is set, we try to set (and/or restore) a null background,
which causes issues. This patch removes the custom backgrounding because it
causes issues (and because `.withLayout().background()` should be equivalent,
minus the extra padding imposed due to double padding on 9-patches).

Reviewed By: sriramramani

Differential Revision: D4880527

fbshipit-source-id: 211679b267e18367ba8d6b3fa12e22242c540a61