Fix weird base context bug in sample app
Summary:
The demo components are created when you start DemoListActivity and stored in a static map.  Each individual demo however, runs in DemoActivity. This means you can get a component with a context based on activity (DemoListActivity) that gets mounted in a view created by a context based on another activity (DemoActivity) and fail ComponentTree#hasSameBaseContext().

This diff makes the easy fix by re-initializing the demos in each activity. That way we insure the components and views they're mounted in have the same base activity.

Reviewed By: emilsjolander

Differential Revision: D4881062

fbshipit-source-id: 5051647854f3bce564b304b0d92cff656eba56b1