Fix tests for litho-widget
Summary:
D4856125 broke the tests by introducing a first PowerMock-enabled test. :(
(Not blaming the author here - it's our tooling that's broken.)

I'm enabling the same options as we have in litho-it, and I plan to move
everything into `litho-it` and find a way to move the special snowflakes that we
have to a module where we can apply those options to while the rest can continue
to run as normal.

Reviewed By: emilsjolander

Differential Revision: D4897345

fbshipit-source-id: 6293edf7835b357c0346c9ef821d3f598b62a501