Split powermock tests into separate target
Summary:
`powermock_srcs` is dead. Long live `it-powermock`!

See the README for a longer explanation. But tl;dr: Make it more obvious where
the slow, brittle stuff lives for both internal and external contributors. This
should hopefully reduce the potential for breakages.

It also removes the separate target for widget tests and kills the
`powermock_srcs` directive altogether so you'll get an error if you try to
reintroduce it.

Reviewed By: muraziz

Differential Revision: D4898361

fbshipit-source-id: ac4a8732f0418a8bae000f94db5ac0c6eaf87603