Reset translation before recording end value of Views layout for animation
Summary:
@public

Before this diff, if we were in a layout animation then translationX/Y might not be 0. If so, the layout for a new end state would incorrectly add that translation to its end state. We now reset the translation before recording the end value. The previous state is restored before the animation continues.

Since View translationX/Y are pretty special in this regard (the fact that you aren't updating the actual value but instead an offset to that value) I don't feel too bad special casing it.

Reviewed By: marco-cova

Differential Revision: D4919143

fbshipit-source-id: aa24c302c970023065b7b69d9cde72df059bc721