Measure sticky header after initializing it
Summary:
We were relying on `RecyclerViewWrapper`'s `onMeasure()` callback to measure sticky header after it has been initialized. However, there are cases when `onMeasure()` of wrapper is not called after we initialized sticky header. For example, if we call `scrollToPosition()` after the `RecyclerView` has been initialized with the data from adapter `onMeasure()` of wrapper is not called, hence the sticky header will end up having the height of 0 even though it has been initialized with component with non-0 height.
Explicitly measuring sticky header every time we initialize it with new component solves the issue.

Reviewed By: marco-cova

Differential Revision: D4924227

fbshipit-source-id: d241814404ee95d5b6faf4346a0831e98adf7c24