Exposing getRecyclerViewWrapper in RecyclerEventsController to show the loading spinner
Summary: As part of the planned viewing redesign of Video Home, one of the features is to show the spinner when the watchlist (section component) is being refreshed. In order to achieve this, getRecyclerViewWrapper method has been exposed in RecyclerEventsController. VideoHomeWatchlistFragment will use the recyclerViewWrapper to show the spinner when requestRefresh is called.

Differential Revision: D4928611

fbshipit-source-id: 9d5c3dc42546cce839bd187131f70b2d2453d1d1