Remove redundant dependencies.
Summary:
The support-v4 library was modularized into distinct components like
support-compat, support-core-ui, support-fragment. We should not depend
on the modules as well as the unmodularized version.

I got errors when running a buck build unrelated to this change so ill need to verify that this compiles with buck. `./gradlew assemble` succeeds.
Closes https://github.com/facebook/litho/pull/127

Reviewed By: muraziz

Differential Revision: D4937802

Pulled By: passy

fbshipit-source-id: 4991d5527434f9c582d9d1975ff8b1e4813160d2