Fix DisplayListDrawable invalidation issue
Summary:
For efficiency reasons we precompute displaylists of `TextDrawable` when we are on main thread and wrap it with `DisplayListDrawable`. Consider the following scenario: `RecyclerView` contains `LithoView` children which in turn contains `Text` components. For the visible items we can precompute display lists for text as they are generated on main thread. `RecyclerBinder` can compute layout for additional items that are not visible on screen on background thread and for those cases the display lists for `Text` components will not be pre-computed. Now when you scroll the list and the items for which layout has been precomputed on background thread appears on screen the `TextDrawable` will be mounted on the host, which will register host (`ComponentHost`) as a callback to `TextDrawable`. Now, you click on a `Text` component which just appeared on the screen and mount update will happen. Now we have display list for that `Text` component and when we update mount state we wrap `TextDrawable` with `DisplayListDrawable` so that we can draw precomputed display lists. When wrapping `TextDrawable` with `DisplayListDrawable` we set `DisplayListDrawable` as a callback to `TextDrawable` which will override its current value which is the `ComponentHost` hosting that text. As we no longer call `ComponentHost.mountDrawable()` on this updated drawable we never set `ComponentHost` as a callback to `DisplayListDrawable` which was causing this issue. When we clicked text the `DisplayListDrawable` had no callback therefore its invalidation was lost and not propagated to its hosting view (`ComponentHost`).

Setting `TextDrawable`'s host to the `DisplayListDrawable` when wrapping solves the issue.

Reviewed By: pasqualeanatriello

Differential Revision: D4946092

fbshipit-source-id: 86d814213ad70af08c2d497a537a009576b088b5