Fix EditText when not using default inputType
Summary:
We really only want to be changing the inputType when we absolutely have to, since it causes the keyboard to become sticky and also resets the selection. Since EditText gets refreshed on every character, we need to make sure that we only update the inputType the first time that the Component is loaded.

`setSingleLine` changes the input type, so we need to make sure that we are comparing against the correct thing.

Reviewed By: mihaelao

Differential Revision: D4953870

fbshipit-source-id: a240a4bf7655ee02628d007f5740df69305d9505