Handle mount item changing during animation
Summary:
There are cases where a transitionKey moves to new position in the hierarchy, resulting in it being mounted to a different mount item (see test plan for an example). This presented two problems:
1. The new mount item was missing the animation state for the transition key (the AnimatedPropertyNodes associated with that transition key)
2. The old mount item has animation values set on it -- we need to reset everything before it gets re-used in the hierarchy

The first is solved by providing a way to update the mount item associated with a AnimatedPropertyNode and then migrating all those AnimatedPropertyNodes to the new mount item.

The second is solved by reseting all properties after current values are recorded in onPreMountItem

Reviewed By: marco-cova

Differential Revision: D4935171

fbshipit-source-id: cf73aac32a4ceeb8a63003ae9eaf27135b6a1db3