fb4a crashes when trying to get to internal settings with aapt2
Summary: At the moment, aapt2 doesn't gracefully handle merging styleable attributes with the same name. Before Google releases new aapt2 version, the easiest workaround would be to not use duplicate styleable attr names. Luckily there is not a lot of such clashes at the moment, so it is a reasonable workaround at this time.

Reviewed By: pasqualeanatriello

Differential Revision: D4945454

fbshipit-source-id: cd352b285ff2e79af69a47bbf2eaa39bd2ef100b