Add CMake build
Summary:
It's building, it's building!

We're not using this yet, but that's coming next. Promise.

This brings all the three parts together: yoga(core), libfb and yogajni. These
are are the parts we need to build the native libraries - natively - with
Gradle and should allow us to remove the dependency pre-published artifacts to
build litho and the samples. Instead, we can directly build from the bundled
source which reduces the overhead that comes with any breaking Yoga changes.

What this isn't is a way to build a *host-native* version of yogajni yet. This
means, we still need to shell out to buck when running the tests. But in theory
this should be possible to fix if we just directly call out to cmake instead of
using the NDK for this.

Reviewed By: IanChilds

Differential Revision: D4905779

fbshipit-source-id: ce82dd41b5d6398bf122ad574303b64cf8dbd34c