Reset animated properties when unmount animations finish
Summary:
When a disappear animation completes, we need to reset all the properties we set on the mount content before telling MountState to clean it up (otherwise these properties will remain set when the mount content is reused).

I had to move changeType to AnimationState since there can be multiple updates (each of which clears all the TransitionDiffs), but we need to keep track of whether we're doing a disappear animation for when the animation finishes.

Reviewed By: marco-cova

Differential Revision: D4969437

fbshipit-source-id: 806fdec358e40ca01d47a513bcf677d815e2bd37