Move current TransitionDiff to AnimationState
Summary:
tl;dr: No expected behavior change, we will now keep track of just AnimationState instead of AnimationState and TransitionDiff in parallel.

Rationale: we were keeping track of TransitionDiffs and AnimationStates in parallel (something emilsjolander asked the reason for in D4956834). I originally thought that was fine, but I needed to start using the changeType when animations end in order to tell whether to reset the mount item: this is located on the transition diff, but actually can persist across multiple animations. Moving this all to AnimationState will simplify things.

Reviewed By: marco-cova

Differential Revision: D4985357

fbshipit-source-id: d5df1d2efadf53633bb0ffc601e34dc8b86835e2