Hook AutoTransitions into MountState to provide animations for feed
Summary:
Depends on D4969549, D4969554

This diff uses the developer define AutoTransitionSet to generate the appropriate dataflow animations at mount time. To do this, we store metadata about what keys are currently mounted in MountState. On ‘dirty’ mounts (e.g. non-incremental mounts), we compare these keys to the transition keys present in the new LayoutState to determine what is appearing/disappearing/changing. Finally, we use that information to create the correct dataflow animations.

Reviewed By: marco-cova

Differential Revision: D4969580

fbshipit-source-id: ea498a9c3fbcff98e2e568af33307e8adcb9be1d