Add fading title bar for new Album consumption.
Summary: Previously tried with adding a blue overlay view besides the section component so that it can be faded while scrolling, however that won't work with landscape view. So changed the approach to the below one: add a tag for the cover photo FbFrescoComponent so that it can be looked up in run time and can be faded while scrolling.

Differential Revision: D4994733

fbshipit-source-id: 65464816321ba701356652fb0b7d5564103f6d1e