Mount everything when setting hasTransientState
Summary: If we have transient state then we want to make sure we have everything mounted so that if the view is animated we don't end up with the view on screen without the correct content. This also means that we should wait until we no longer have transient state before mounting/unmounting anything else.

Reviewed By: marco-cova

Differential Revision: D4994810

fbshipit-source-id: 50e22a958f62ad1f60042db11631dad80b1205d4