Prefetch display lists. Part 1. Pass information about whether to prefetch from RecyclerBinder down to LayoutState.
Summary:
Display list prefetching consists of two operations:
- Collecting `LayoutState`/`LayoutOutputs` for further processing (creating display lists from them) asynchronously.
- Trigger `Runnable` when scrolling scrollable content to actually process items from previous operation.

They are enabled separately, but we need to have a way to enable/disable them from single place. In our case that single place is `RecyclerBinder` but it should be easy to extend it to other types of adapters/widgets as long as that adapter passes right flag when creating `ComponentTree` and calls `DisplayListPrefetcherUtils.prefetchDisplayLists()` when scrolling.

In the first part of this stacked diff adding capability to propagate boolean flag (`canPrefetchDisplayLists`) from `RecyclerBinder` to each of its individual child `ComponentTree`'s `LayoutState`. Whenever it's scrolled we are triggering  `DisplayListPrefetcherUtils.prefetchDisplayLists()` which will eventually post runnable (that part will be added in further diffs) that will process collected `LayoutState`/`LayoutOutput`s.

Reviewed By: pasqualeanatriello

Differential Revision: D4987475

fbshipit-source-id: 17b463023c991852505a9edeca25195ff3554467