Only prepare and execute transitions on dirty mounts
Summary:
There was inconsistency with when we'd prepare a transition and when we'd execute it: now they are the exact same for both, and only do it on dirty mounts.

This fixes a bug where when you tapped like, then scrolled away, then came back, you'd see a partially animated layout.

Note, this doesn't work for incremental mount, but we currently have that disabled anyway.

Reviewed By: marco-cova

Differential Revision: D5021079

fbshipit-source-id: 282d1108e89dea8b61d5e5600582308f8e762f92