Only re-use display lists on main thread
Summary: We saw a bug where if you re-use a display list on a background thread (as could previously happen in `collectResults`, then we wouldn't actually update the display list in `collectDisplayLists`, since that only gets called if you're on the main thread. Therefore we'd be using an old displaylist, and if the bounds should have been updated (e.g. if the new text is wrapped in a view and the old text was not) then the text wouldn't get printed correctly.

Reviewed By: muraziz

Differential Revision: D5028182

fbshipit-source-id: 18c12d7cc928c32ac73d13a1138311636c1adc56