Explicitly enable Display List creation for android versions when new versions come out
Summary: For different versions of Android we have different implementation classes of DisplayList. And so far we had a condition which said that if version is above certain API number use particular implementation that is specific to that API number. The potential issue here is that when new version is released and DisplayList API changes, then we need to create a new implementation class, however, in the meantime the app using this feature will crash. From this diff onwards we will explicitly enable display list creations for new versions after testing them.

Reviewed By: pasqualeanatriello

Differential Revision: D5020100

fbshipit-source-id: bf8383200e039979285821faadbc348eefaa9ae5