Support only clipping top corners in CardClip
Summary:
I have a card view which is specced to have rounded corners - but I only want them on the top-left and top-right corners. The reasoning is that the card is vertically scrollable and having rounded corners on the bottom might make the user think it's not scrollable (i.e. it looks like the whole card is in the viewport when it isn't) - see the test plan for a screenshot.

CardClip mostly works for this use-case except for that issue so the proposal in this diff is to add a parameter to turn off rounding bottom corners. This could be generalized to control each corner individually if someone ever wanted that.

Let me know if you'd prefer not to have this in the library - I can implement my own clipping drawable if that's the case.

Reviewed By: muraziz, IanChilds

Differential Revision: D5025370

fbshipit-source-id: 2857ab8cb0fad64946fc2c492e873198c2c96007