ImageSpec.shouldUpdate just uses pointer equality
Summary:
We don't want to end up reusing drawables in more than one place. Consider the following:

Image 1 is created with Drawable A, and mounted in View V1
Image 2 is created with Drawable B, and mounted in View V2

For some reason, Image 2 is set to View V1, and because Drawable A and Drawable B are equal, shouldUpdate returns false, and Drawable A remains in View V1 (but View V1 is showing Image 2).

Image 1 is given a new View V3, and Drawable A is also mounted in View V3. Drawable A is now used in two places and any bounds changes will affect both of the places.

Instead we can just reuse a drawable if it is the same prop.

Reviewed By: pasqualeanatriello

Differential Revision: D5052364

fbshipit-source-id: 4cf80129650a8d6fe112b614c00a6e64b3003dab