Cope with null annotations on TypeElements
Summary: Bit mixed on this. We need it to make the tests work with gradle - we want to be able to create a spec model without having the annotation (so that we can create spec models in test scenarios without running validation over them). However, it looks a bit strange that you could have a null annotation.

Reviewed By: passy

Differential Revision: D5060788

fbshipit-source-id: d230a2cb54539434aaf76092417d596f477cda48