Add onInterceptTouchEvent
Summary:
This is a new API to match the Android API ViewGroup.onInterceptTouchEvent. It should work in the same way - you set an `interceptTouchEvent` handler to your layout and it is invoked when intercept touch event would be invoked.

There is a hole in the implementation right now - since you can't add a listener in the same way you can for clicks and touches, this only works on ComponentHosts (where we can override the method) but doesn't work for Mount Specs. I think that's probably okay, but should we crash in MountState to make that clear, or what do people think?

Reviewed By: marco-cova

Differential Revision: D5069090

fbshipit-source-id: e3de8b2d44ecdf32c97ac1f3f1a10e825d21e874